CA fixes
- Code Analysis fixes
- TODO list fixes
- More function definition ordering

+++ galaxy/quasar/src/qs/texture/Texture.cpp
@@ -70,7 +70,7 @@ namespace qs
 		glBindTexture(GL_TEXTURE_2D, m_texture);
 
 		stbi_set_flip_vertically_on_load(true);
-		unsigned char* data = stbi_load_from_memory(buffer.data(), buffer.size_bytes(), &m_width, &m_height, nullptr, STBI_rgb_alpha);
+		unsigned char* data = stbi_load_from_memory(buffer.data(), static_cast<int>(buffer.size_bytes()), &m_width, &m_height, nullptr, STBI_rgb_alpha);
 
 		if (data)
 		{
